Added BaseGeoWidgetTest to remove duplicated code
