+++ collect_app/src/main/java/org/odk/collect/android/widgets/BaseGeoWidget.java
@@ -11,8 +11,8 @@ import org.odk.collect.android.listeners.PermissionListener;
 import org.odk.collect.android.widgets.interfaces.GeoWidget;
 
 public abstract class BaseGeoWidget extends QuestionWidget implements GeoWidget {
-    Button startGeoButton;
-    TextView answerDisplay;
+    public Button startGeoButton;
+    public TextView answerDisplay;
     protected boolean readOnly;
 
     public BaseGeoWidget(Context context, QuestionDetails questionDetails) {
